requirements.txt
Flask
Flask-SQLAlchemy==3.0.2
Werkzeug==2.2.2
Flask-Limiter==2.4.0
openai
anthropic
#google-cloud-aiplatform==1.16.1
python-dotenv==0.19.2
gunicorn==20.1.0